searchFunction
