+++ app/src/main/java/com/mealmastercookingrecipesapp/Controller/FileHandler.java
@@ -1,4 +1,4 @@
-package com.mealmastercookingrecipesapp;
+package com.mealmastercookingrecipesapp.Controller;
 
 import android.content.Context;
