+++ app/src/main/java/com/mealmastercookingrecipesapp/Controller/FragmentManager.java
@@ -1,21 +1,29 @@
-package com.mealmastercookingrecipesapp;
+package com.mealmastercookingrecipesapp.Controller;
 
 import android.content.Context;
 import android.content.res.Resources;
+import android.graphics.PorterDuff;
 import android.graphics.Typeface;
+import android.view.Gravity;
 import android.view.View;
+import android.widget.FrameLayout;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.core.content.ContextCompat;
+
+import com.mealmastercookingrecipesapp.Controller.FavoriteManager;
+import com.mealmastercookingrecipesapp.R;
 import com.squareup.picasso.Picasso;
 
 import jp.wasabeef.picasso.transformations.RoundedCornersTransformation;
-
+// View view, Context context, Resources resources, boolean isSearchFragment)
 public class FragmentManager {
-    public static void addImageToFragment(String title, String imageUrl, String id, View view, Context context, Resources resources, boolean isSearchFragment) {
+    public static void addImageToFragment(String title, String imageUrl, String id, View view, int iconId, Context context, Resources resources, boolean isSearchFragment) {
         // Erstelle einen TextView für den Titel des Bildes
+        FavoriteManager favoriteManager = new FavoriteManager(context);
         TextView textView = new TextView(context);
         textView.setText(title);
         textView.setTypeface(null, Typeface.BOLD); // Fett anzeigen
@@ -23,6 +31,8 @@ public class FragmentManager {
         textView.setTextColor(resources.getColor(R.color.black));
         // Erstelle ein neues ImageView-Element
         ImageView imageView = new ImageView(context);
+        //Erstelle final copy of ImageView
+        final ImageView finalImageView = imageView;
         // Definiere die gewünschte Zielgröße für das Bild
         int targetWidth = resources.getDisplayMetrics().widthPixels; // Breite des Bildschirms
         int targetHeight = (int) (targetWidth * 2); // Verhältnis der Höhe zur Breite (hier: 0.75)
@@ -70,11 +80,68 @@ public class FragmentManager {
         textLayoutParams.setMargins(margin,0,0,0);
         textView.setLayoutParams(textLayoutParams);
 
-        // Füge das ImageView und den TextView zum LinearLayout hinzu
-        linearLayout.addView(textView);
-        linearLayout.addView(imageView, imageLayoutParams);
+        // Füge das Icon hinzu
+        ImageView iconView = new ImageView(context);
+        iconView.setImageResource(iconId);
 
-        // Füge das LinearLayout mit dem Bild und Titel zum Layout des Fragments hinzu
+        if (favoriteManager.isFavorite(id)) {
+            iconView.setColorFilter(ContextCompat.getColor(context, R.color.pink), PorterDuff.Mode.SRC_IN);
+        } else {
+            iconView.setColorFilter(ContextCompat.getColor(context, R.color.grey), PorterDuff.Mode.SRC_IN);
+        }
+
+        //final boolean[] isFavorised = {false};
+        iconView.setOnClickListener(new View.OnClickListener(){
+            @Override
+            public void onClick(View v){
+
+                String imageViewId = (String) finalImageView.getTag();
+
+                if (favoriteManager.isFavorite(imageViewId)) {
+                    iconView.setColorFilter(ContextCompat.getColor(context, R.color.grey), PorterDuff.Mode.SRC_IN);
+                    //zur sicherheit check, ob schon de-favorisiert
+                    //entferne von Arrayliste der favorite
+                    favoriteManager.removeFromFavorites(imageViewId);
+                    Toast.makeText(context, "Removed ID: " + imageViewId, Toast.LENGTH_SHORT).show();
+
+                } else {
+                    iconView.setColorFilter(ContextCompat.getColor(context, R.color.pink), PorterDuff.Mode.SRC_IN);
+                    //zur sicherheit check, ob schon favorisiert
+                    //füge zur ArrayListe der favoriten hinzu
+                    favoriteManager.addToFavorites(imageViewId);
+                    Toast.makeText(context, "Added ID: " + imageViewId, Toast.LENGTH_SHORT).show();
+                }
+            }
+        });
+        //Um die größe des Icons einzustellen
+        int iconSize = resources.getDimensionPixelSize(R.dimen.icon_size);
+
+        // Konfiguriere die Layout-Parameter für den IconView
+        FrameLayout.LayoutParams iconLayoutParams = new FrameLayout.LayoutParams(
+                iconSize,
+                iconSize
+        );
+
+        //Positioniere den Icon oben rechts von ImageView
+        iconLayoutParams.gravity = Gravity.TOP | Gravity.END;
+
+        //Fügt ein Margin für das IconView hinzu
+        int iconMargin = resources.getDimensionPixelSize(R.dimen.icon_margin);
+        int iconMarginRight = resources.getDimensionPixelSize(R.dimen.icon_margin_right);
+        iconLayoutParams.setMargins(0, iconMargin, iconMarginRight, 0);
+
+        iconView.setLayoutParams(iconLayoutParams);
+
+        FrameLayout frameLayout = new FrameLayout(context);
+
+        //Füge das Imageview und IconView zum FrameLayout hinzu um sie übereinander zu rendern
+        frameLayout.addView(imageView, imageLayoutParams);
+        frameLayout.addView(iconView, iconLayoutParams);
+
+        // Füge den TextView und das FrameLayout zum LinearLayout hinzu
+        linearLayout.addView(textView);
+        linearLayout.addView(frameLayout);
+
         if (isSearchFragment){
             LinearLayout fragmentLayout = view.findViewById(R.id.searchFragmentLinearLayoutRecipes);
             fragmentLayout.addView(linearLayout);
