+++ app/src/main/java/com/mealmastercookingrecipesapp/MainActivity.java
@@ -6,19 +6,13 @@ import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 
 import android.os.Bundle;
-import android.util.Log;
-import android.view.View;
-import android.widget.Button;
-import android.widget.TextView;
 
-import com.android.volley.Request;
-import com.android.volley.Response;
-import com.android.volley.VolleyError;
-import com.android.volley.toolbox.JsonObjectRequest;
-import com.android.volley.toolbox.Volley;
+import com.mealmastercookingrecipesapp.Controller.ApiHandler;
+import com.mealmastercookingrecipesapp.View.FavoriteFragment;
+import com.mealmastercookingrecipesapp.View.HomeFragment;
+import com.mealmastercookingrecipesapp.View.MenuFragment;
+import com.mealmastercookingrecipesapp.View.SearchFragment;
 import com.mealmastercookingrecipesapp.databinding.ActivityMainBinding;
-
-import org.json.JSONObject;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -26,7 +20,7 @@ public class MainActivity extends AppCompatActivity {
     ApiHandler apiHandler;
     String url = "https://api.spoonacular.com/recipes/random?number=1&apiKey=";
 
-    String apiKey = "2a23994b963346b292d20fbc9dd02e5d";
+    String apiKey = "8901d0f1e4ac4e18b88b126d663d62bd";
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
