+++ app/src/main/java/com/mealmastercookingrecipesapp/Model/Recipe.java
@@ -1,4 +1,4 @@
-package com.mealmastercookingrecipesapp;
+package com.mealmastercookingrecipesapp.Model;
 
 public class Recipe {
