+++ app/src/main/java/com/mealmastercookingrecipesapp/Model/RecipeCallback.java
@@ -1,4 +1,6 @@
-package com.mealmastercookingrecipesapp;
+package com.mealmastercookingrecipesapp.Model;
+
+import com.mealmastercookingrecipesapp.Model.Recipe;
 
 public interface RecipeCallback {
     void onSuccess(Recipe recipe);
