+++ app/src/main/java/com/mealmastercookingrecipesapp/Model/RecipeCallbackArray.java
@@ -1,4 +1,6 @@
-package com.mealmastercookingrecipesapp;
+package com.mealmastercookingrecipesapp.Model;
+
+import com.mealmastercookingrecipesapp.Model.Recipe;
 
 public interface RecipeCallbackArray {
     void onSuccess(Recipe[] recipes); // Array von Rezepten
