+++ app/src/main/res/layout/fragment_favorite.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".FavoriteFragment">
+    tools:context=".View.FavoriteFragment">
 
     <ScrollView
         android:layout_width="match_parent"
