+++ app/src/main/res/layout/fragment_home.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".HomeFragment">
+    tools:context=".View.HomeFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
