+++ app/src/main/res/layout/fragment_menu.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FFFFFF"
-    tools:context=".MenuFragment">
+    tools:context=".View.MenuFragment">
 
     <ScrollView
         android:layout_width="match_parent"
