+++ app/src/main/res/layout/fragment_search.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@color/white"
     android:orientation="vertical"
-    tools:context=".SearchFragment">
+    tools:context=".View.SearchFragment">
 
     <TextView
         android:id="@+id/textView"
